Update and rename Pouch_of_tiles.py to pouch.py
